+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Create link for creating Gitlab personal access token with required scopes from Gitlab API url.
 
 ## [1.8.0] - 2023-04-23